Fix UFix maxValue bug from issue #343.
This commit fixes a copy/paste bug in the `maxValue` calculation for the
`UFix` type in `FixedPoint.scala`.

The original `maxValue` calculation was copied from `SFix`, and resulted
in spurious assertion failures for users who were trying to assign the
maximum possible value to their `UFix` instances.

The issue was fixed by removing the `- 1` in the calculation that was
being used to account for the sign bits in the SFix calculation.