Axi4Decoder: fix null pointer in Axi4WriteOnlyDecoder
In Axi4WriteOnlyDecoder, when the aggregated slave address mapping
is smaller than what the address bits can cover, the errorSlave
variable will be set, otherwise it is NULL.

The original code does not check whether this errorSlave is NULL
or not. Therefore, in some cases, the code will not be able to compile.