feat: protocol simulation API; implemented for AHB
The API is not protocol-dependent and can be implemented for other
protocols as well (even Stream and Flow I guess, will have to try ;).

Notice how the way AHB is implemented matches the ARM specification.
Same for its tests.

This is WIP, the API is not stabilized yet (aka function names can still
change). Also, it is not possible to check if it works because other
parts of the code do not build. (It was working before being integrated,
I swear ;)

Discussion to have / things to change:

* What about AhbAttributes for master behavior? Should I use it for all
  transactions or there is a better way to do this?
* Slave "setOn" functions should be renamed to explicitly manage 3 types
  of arguments: reader (addr => data) and writer (addr, data => Unit);
  then Behavior (Req => Resp) and finally Builder (Req => () => resp,
  aka Req => Handler). Then implicit conversion Behavior => Builder
  should be removed.
* Should I add a class to just read signals and check that all "must" of
  the spec are applied or is it already a thing?
* AhbMaster is not implemented yet (it is almost the same, just one
  signal changes I think)
* Parts of
  tester/src/test/scala/spinal/tester/scalatest/lib/bus/amba3/ahblite/sim/hardware.scala
  are already implemented somewhere I guess.