UartCtrlRx: bitTimer.reset() to center bit in Majority Voter
We don't just want to find the center of the bit.

We want the center of the bit, when it is in the center of the
Majority Voter.  So need to allow extra sampler.tick counts for propagation
by adding again samplingSize/2 to account for half the width of the
Majority Voter.