Remove pinned C++ version for verilator
Fixed C++11 causes issues with latest verilator. Instead
of requiring C++11, go back to before #278 where the default
version of the compiler was used.

The override was added to support seemingly older versions of
gcc, to keep that support provide an environment variable
that allows the setting of arbitrary flags. This also
allows, e.g. the use of an alternative linker.