formal: add formalXXX variants of past/rose/fell
In commit #9f86e95 the generation of past, rose, fell, changed and
stable functions was changed. Instead of using the systemverilog
language specific builtin functions $past, $rose, etc. their
behaviour was modelled directly within spinal with `RegNext`. This adds
the functions back with the explicit formalXXX for those that want to
use it.