formal: disable automatic cover directive for asserts
GHDL automatically generates a cover directive for the precondition of
asserts. This is generally a good thing as we want to know if the assert
could ever be triggered at all. This is not identical behaviour to the
default verilog SymbiYosys backend though and disabled intentionally.

To manually get the same effect add an additional cover directive with
the assert precondition:
```scala
when(precondition) { assert(condition) }
cover(precondition)
```