add code generator for the blackbox ROM model
This generator is slightly less than ideal because the code will
still contain some inactive code that creates a ROM variable and
initializes it with a .bin file even if a blackbox is used.

The reason is that the `mem` component itself doesn't have a way
to know that it's a black box (as far as I can tell), so the safe
thing to do is to generate both artifacts.

Thus, you end up with two things:

- a *.bin file that is suitable for a $readmemb directive in the case
that no blackboxing is requested

- a .v file that is suitable for blackobxing

The .v file implements the ROM as a `case` statement so that the
blackbox can simulate with tools that don't support $readmemb directives
(such as some ASIC flows).