Add SimEquiv trait and implicit conversion
The SimEquiv trait represents the relationship where a scala type can
correspond to a SpinalHDL type during simulation. This allows methods
which assign or get simulation values to be written in terms of
generics, thus possibly greatly reducing code repetition.

More concretely, it allows you to use #= between a SpinalHDL Vec and
Scala List without the need to implement it for every type combination.