Mask tilelink bus valid bit when the bus is marked as assignDontCare
When the tilelink bus is marked as assignDontCare, it will
be assigned as random bits. This will break the system if
the valid bit is assigned as true and the channel will send
an invalid request.

Mask the valid bit of the channel A and B if the bus is a
master port.

Signed-off-by: Inochi Amaoto <inochiama@gmail.com>